motor_voltage is now a double
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@4230 f5854215-dd47-0410-b2c4-cdd35faa7885

+++ mechanism/hardware_interface/include/hardware_interface/hardware_interface.h
@@ -70,7 +70,7 @@ public:
   double last_commanded_effort_;
   double last_measured_effort_;
 
-  int motor_voltage_;
+  double motor_voltage_;
 
   int num_encoder_errors_;
   int num_communication_errors_;
